pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel>=0.41.0"]
build-backend = "setuptools.build_meta"

[project]
name = "migrationiq"
version = "0.1.0"
description = "Git-aware migration safety CLI for Django and Alembic projects"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "MigrationIQ Contributors"},
]
keywords = ["django", "alembic", "migrations", "safety", "cli", "git"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Framework :: Django",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Database",
    "Typing :: Typed",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "sqlalchemy>=2.0.0",
    "sqlparse>=0.5.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
migrationiq = "migrationiq.cli.main:app"

[project.urls]
Homepage = "https://github.com/migrationiq/migrationiq"
Repository = "https://github.com/migrationiq/migrationiq"
Issues = "https://github.com/migrationiq/migrationiq/issues"

[tool.setuptools.packages.find]
include = ["migrationiq*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.coverage.run]
source = ["migrationiq"]
omit = ["tests/*"]

[tool.coverage.report]
show_missing = true
fail_under = 80
